refactor notifications to work within the App and CreateSource components
