+++ ui/src/shared/actions/notifications.js
@@ -1,4 +1,10 @@
 export function publishNotification(type, message) {
+  // this validator is purely for development purposes. It might make sense to move this to a middleware.
+  const validTypes = ['error', 'success', 'warning']
+  if (!validTypes.includes(type) || message === undefined) {
+    console.error("handleNotification must have a valid type and text") // eslint-disable-line no-console
+  }
+
   return {
     type: 'NOTIFICATION_RECEIVED',
     payload: {
